src/main/java/com/example/demo/DemoApplication.java
package com.example.demo;

import lombok.extern.slf4j.Slf4j;
import org.springframework.boot.CommandLineRunner;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.context.annotation.Bean;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@Slf4j
@RestController
@SpringBootApplication
public class DemoApplication {

    public static void main(String[] args) {
        SpringApplication.run(DemoApplication.class, args);
    }

    @GetMapping
    public String welcome() {
        return "welcome";
    }

    @GetMapping("/hi")
    public String hello() {
        return "hello";
    }

    @Bean
    public CommandLineRunner runner() {
        return args -> log.info("hello from runner");
    }
}
